revert: removes baseContext and subscribeBaseContext and stripped the subscribeContext from extra methods, also reverted the executeContext to the previous state
