+++ theoriq/agent.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import os
-from datetime import datetime
 from typing import Any, Dict, Optional
 
 import biscuit_auth
