+++ theoriq/biscuit/authentication_biscuit.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime, timedelta, timezone
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
 from biscuit_auth import Biscuit, PrivateKey
